- modularizando o código.

+++ src/back/LowerBound.java
@@ -5,7 +5,7 @@ package back;
  */
 public class LowerBound implements Estimator {
 
-    public int estimate (int collision_slots, int successful_slots){
-        return collision_slots*2+successful_slots;
+    public int estimate (double collision_slots, double successful_slots){
+        return (int) (collision_slots*2+successful_slots);
     }
 }
